backend/pyproject.toml
[project]
name = "ag-ui-sql-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Lennart Weyrowitz",email = "lennart.weyrowitz@accesify.de"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "cryptography (>=45.0.5,<46.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pymssql (>=2.3.7,<3.0.0)",
    "plotly (>=6.3.0,<7.0.0)",
    "redis-om (>=0.3.5,<0.4.0)",
    "pydantic-ai (>=1.0.16,<2.0.0)",
    "logfire[fastapi] (>=4.12.0,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"